[add] 인터넷 연결이 되어 있을 때 서버에 올라가지 않았던 데이터 업로드
- repository 통해서 load 시 null이 반환되는 오류 남아 있음, 이유는 아직

+++ WiFiProjectBackground/app/src/main/java/wifiproject/background/serverconnection/PushResultModel.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @Setter
 public class PushResultModel {
     @SerializedName("count")
-    Integer count;
+    Long count;
 }
